Delete the filters DLL
Per David email dec 5 2022:

At some point I made a stand-alone DLL with that only do filtering. I created that to have more tuning parameters in a general filter DLL instead of extending the hard-coded ones in other DLLs. I suspect it ended up in some demo model and kept being shared while not being used anywhere.

I suggest to remove it from the Github repo if it is not being used in the model.